auth responses fixed for mobile

+++ src/main/java/com/example/SynClock/model/DTOs/CreateAlarmClockDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CreateAlarmClockDTO {
     private String time;
-    private Long groupId;
+    private Integer groupId;
 }
